# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Navidrome Music Room Gateway API
  version: 1.1.0-dev
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
  description: |
    Versioned room coordination API for MusicMate. Media bytes never pass through this API;
    clients stream audio, artwork, and lyrics directly from Navidrome with their own credentials.
servers:
  - url: https://music.example.com/music-room/api/v1
security:
  - roomSession: []
...
paths:
...
  /auth/exchange:
    post:
      tags: [Auth]
      summary: Exchange an OpenSubsonic salt/token proof for a 15-minute room session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthProof'
      responses:
        '200':
          description: Authenticated room session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthExchange'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /auth/session:
    delete:
      tags: [Auth]
      summary: Revoke the current in-memory room session
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Error' }
  /rooms:
    get:
      tags: [Rooms]
      summary: List rooms visible to the current user
      responses:
        '200':
          description: Rooms where the user is a member; administrators see all rooms
          content:
            application/json:
              schema:
                type: object
                required: [rooms]
                properties:
                  rooms:
                    type: array
                    items: { $ref: '#/components/schemas/Room' }
    post:
      tags: [Rooms]
      summary: Create a room
      description: Requires both Navidrome adminRole and plugin authorization.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreate'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '403': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    get:
      tags: [Rooms]
      summary: Read one room
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Rooms]
      summary: Update a room as owner or administrator
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomPatch' }
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Rooms]
      summary: Permanently delete a room as owner or administrator
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '204': { description: Room deleted }
        '403': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/close:
    post:
      tags: [Rooms]
      summary: Close a room to new joins and queue changes
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Closed room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '403': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/reopen:
    post:
      tags: [Rooms]
      summary: Reopen a room when the plugin lease is healthy
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Reopened room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '503': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/invites:
    parameters: [{ $ref: '#/components/parameters/RoomID' }]
    get:
      tags: [Invitations]
      summary: List invitation metadata without secrets
      responses:
        '200':
          description: Invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items: { $ref: '#/components/schemas/Invite' }
    post:
      tags: [Invitations]
      summary: Create a revocable invitation
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 80 }
                expiresAt: { type: string, format: date-time }
                maxUses: { type: integer, minimum: 1, maximum: 10000, default: 20 }
                singleUse: { type: boolean, default: false }
      responses:
        '201':
          description: Secret is returned once; only its SHA-256 digest is persisted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InviteWithSecret' }
  /rooms/{room_id}/invites/{invite_id}:
    delete:
      tags: [Invitations]
      summary: Revoke an invitation without removing existing members
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - name: invite_id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Identifier' }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Invitation revoked }
        '404': { $ref: '#/components/responses/Error' }
  /invites/redeem:
    post:
      tags: [Invitations]
      summary: Redeem an invitation into persistent room membership
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomID, invite]
              properties:
                roomID: { $ref: '#/components/schemas/Identifier' }
                invite: { type: string, minLength: 32, maxLength: 128 }
      responses:
        '200':
          description: Membership created or already active
          content:
            application/json:
              schema:
                type: object
                properties:
                  member: { $ref: '#/components/schemas/Member' }
                  room: { $ref: '#/components/schemas/Room' }
        '403': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/members:
    get:
      tags: [Rooms]
      summary: List persistent members with current presence
      parameters: [{ $ref: '#/components/parameters/RoomID' }]
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
  /rooms/{room_id}/members/{username}:
    delete:
      tags: [Rooms]
      summary: Remove a member and disconnect their room WebSockets
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - name: username
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Member removed }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/join:
    post:
      tags: [Rooms]
      summary: Enter a room after membership and library checks
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Joined }
        '403': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/leave:
    post:
      tags: [Rooms]
      summary: Deactivate the current membership
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Left room }
        '409': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/snapshot:
    get:
      tags: [Rooms]
      summary: Read the authoritative room snapshot
      parameters: [{ $ref: '#/components/parameters/RoomID' }]
      responses:
        '200':
          description: Snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snapshot' }
  /rooms/{room_id}/history:
    get:
      tags: [Playback]
      summary: Read playback history
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200': { description: History page }
  /rooms/{room_id}/playback:
    post:
      tags: [Playback]
      summary: Change authoritative playback as owner or administrator
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command, expectedRevision]
              properties:
                command: { type: string, enum: [play, pause, seek, next, stop] }
                expectedRevision: { type: integer, format: int64, minimum: 0 }
                positionSeconds: { type: number, minimum: 0 }
      responses:
        '200':
          description: New playback state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playback' }
        '409':
          description: Includes `details.latest` when code is `revision_conflict`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /rooms/{room_id}/queue/tracks:
    post:
      tags: [Playback]
      summary: Add an accessible Navidrome track to the member's pending queue
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track]
              properties:
                track: { $ref: '#/components/schemas/TrackRef' }
      responses:
        '201':
          description: Queue entry accepted; `playback` is present when the first pending track became the paused current track
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueMutation' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/queue/{queue_id}:
    delete:
      tags: [Playback]
      summary: Remove one's own pending entry, or any entry as a manager
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - name: queue_id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Identifier' }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Queue entry removed }
        '403': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/queue/order:
    put:
      tags: [Playback]
      summary: Reorder the complete queue as a manager
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [queueIDs]
              properties:
                queueIDs:
                  type: array
                  maxItems: 1000
                  items: { $ref: '#/components/schemas/Identifier' }
      responses:
        '200': { description: Reordered queue }
        '409': { $ref: '#/components/responses/Error' }
  /rooms/{room_id}/ws-ticket:
    post:
      tags: [Rooms]
      summary: Create a one-use WebSocket ticket valid for 60 seconds
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Ticket and WebSocket URL
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expiresAt, webSocketURL]
                properties:
                  ticket: { type: string }
                  expiresAt: { type: string, format: date-time }
                  webSocketURL: { type: string, format: uri }
  /rooms/{room_id}/ws:
    get:
      tags: [Rooms]
      summary: Upgrade to the room WebSocket
      description: |
        The `ticket` query value is one-use and expires after 60 seconds. Server events are
        `snapshot`, `playback`, `queue`, `history`, `presence`, `room_updated`, and `error`.
      security: []
      parameters:
        - $ref: '#/components/parameters/RoomID'
        - { name: ticket, in: query, required: true, schema: { type: string } }
      responses:
        '101': { description: WebSocket upgraded }
        '401': { $ref: '#/components/responses/Error' }
...
  /admin/diagnostics:
    get:
      tags: [Updates]
      summary: Explicitly export an automatically redacted diagnostic bundle
      description: Returns aggregate runtime, plugin heartbeat, update, and database health only; no URLs, names, room identifiers, invitations, credentials, or row data are included.
      responses:
        '200':
          description: Private JSON attachment generated on demand
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiagnosticBundle' }
        '403': { $ref: '#/components/responses/Error' }
...
  /admin/updates:
    get:
      tags: [Updates]
      summary: Read local updater state as a Navidrome administrator
      responses:
        '200':
          description: Update state
          content:
            application/json:
              schema:
                type: object
                required: [status, repository, managedByLauncher, actions]
                properties:
                  status: { $ref: '#/components/schemas/UpdateStatus' }
                  repository: { type: string }
                  managedByLauncher: { type: boolean }
                  actions: { type: array, items: { type: string, enum: [check, stage, install, rollback] } }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [Updates]
      summary: Check, stage, install, or roll back a signed GitHub release
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [check, stage, install, rollback] }
                version: { type: string }
      responses:
        '200': { description: Check completed }
        '202': { description: 'Stage, activation, or rollback accepted' }
        '409': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    roomSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-15-minute-room-session
  parameters:
    RoomID:
      name: room_id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Identifier' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        A caller-generated key retained in memory for 15 minutes. Repeating the same method,
        path, body, current authorized user, and key replays the original response and adds
        `Idempotency-Replayed: true`; reusing it for a different request returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Identifier:
      type: string
      pattern: '^[A-Za-z0-9_-]{16,128}$'
    AuthProof:
      type: object
      additionalProperties: false
      required: [username, salt, token]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        salt: { type: string, pattern: '^[A-Fa-f0-9]{8,64}$' }
        token: { type: string, pattern: '^[A-Fa-f0-9]{32}$' }
...
    AuthExchange:
      type: object
      required: [sessionToken, expiresAt, user, navidromeBaseURL, gatewayBaseURL, pluginLeaseExpiresAt]
      properties:
        sessionToken: { type: string }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
        navidromeBaseURL: { type: string, format: uri }
        gatewayBaseURL: { type: string, format: uri }
        pluginLeaseExpiresAt: { type: string, format: date-time }
    RoomCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        queueLimit: { type: integer, minimum: 1, maximum: 100, default: 20 }
        playbackMode: { type: string, enum: [fifo, fair_random], default: fifo }
        musicFolderIDs:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1 }
        preloadNextTrack: { type: boolean, default: true }
...
    Room:
      type: object
      required: [roomID, name, ownerUsername, ownerDisplayName, status, queueLimit, playbackMode, musicFolderIDs, preloadNextTrack, capabilities, onlineCount, createdAt, updatedAt]
      properties:
        roomID: { $ref: '#/components/schemas/Identifier' }
        name: { type: string }
        ownerUsername: { type: string }
        ownerDisplayName: { type: string }
        status: { type: string, enum: [open, closed] }
        queueLimit: { type: integer }
        playbackMode: { type: string, enum: [fifo, fair_random] }
        musicFolderIDs:
          type: array
          items: { type: integer }
        preloadNextTrack: { type: boolean }
        capabilities: { $ref: '#/components/schemas/Capabilities' }
        onlineCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...